Fix clippy warnings: derive Default and add #[allow(clippy::derivable_impls)]
- Add #[derive(Default)] to enums and structs where all fields have default values:
  - ThemeMode, Language, State, Margin, RenderPlan, TouchState, ResourceSummary, LightLevels, Declaration
- Add #[allow(clippy::derivable_impls)] to manual Default impls that use:
  - Runtime values (Local::now(), theme::foreground(), CURRENT_DEVICE.dpi)
  - Non-default constants (Vec::with_capacity, specific numeric values)
  - Self::new() calls
  - unsafe { mem::zeroed() }
- Remove empty lines after outer attributes to fix clippy warnings
- Fix ThemeSettings struct definition by removing misplaced attribute

This reduces clippy errors from 278 to 0 on the ARM target.

+++ crates/core/src/view/about_dialog.rs
@@ -73,7 +73,7 @@ impl AboutDialog {
         children.push(Box::new(label));
 
         // Add OK button
-        let button_width = font.plan(LABEL_OK, Some(max_button_width), None).width as i32 + padding;
+        let button_width = font.plan(LABEL_OK, Some(max_button_width), None).width + padding;
         let rect_ok = rect![
             (rect.min.x + rect.max.x - button_width) / 2,
             rect.max.y - button_height - padding,
